+++ .dialyzer_ignore.exs
@@ -1,5 +1,0 @@
-[
-  # Add any Dialyzer warnings to ignore here
-  # Format: {file, line, warning_pattern}
-  # Example: {"lib/some_file.ex", :unknown, :pattern_match}
-]